flask/requirements-no-version.txt
flask
flask_sqlalchemy
flake8
python-dotenv
gunicorn
pyyaml